Simplify report data validation
